+++ changelog.md
@@ -5,6 +5,8 @@
 > [!IMPORTANT]
 > This version is not released yet and is under active development.
 
+- Replace unmaintained `deepmerge` dependency with `mergedeep`.
+
 ## [5.1.1 (2025-08-24)](https://github.com/kdeldycke/click-extra/compare/v5.1.0...v5.1.1)
 
 - Relax Click dependency to account for yanked release. Closes {issue}`1307`.